Concatenate body childNodes
It will be fragile to match exact strings to be removed from a saved
DOMDocument. It's already been observed that this introduces extraneous
newlines and other document normalization could change the strings we
expect to match.

Instead of matching some HTML strings to remove, concatenate the
children of the document body element in order to produce the expected
HTML.

Noted in https://github.com/WordPress/gutenberg/pull/25028#discussion_r483843280